This function could be in utils.py
